:rotating_light: fix warnings

+++ test/src/unit-msgpack.cpp
@@ -446,7 +446,7 @@ TEST_CASE("MessagePack")
 
                 SECTION("-32768..-129 (int 16)")
                 {
-                    for (int16_t i = -32768; i <= -129; ++i)
+                    for (int16_t i = -32768; i <= int16_t(-129); ++i)
                     {
                         CAPTURE(i)
